+++ scripts/knn2img.py
@@ -54,7 +54,7 @@ def load_model_from_config(config, ckpt, verbose=False):
         print(u)
 
     model.cuda()
-    model.eval()
+    model.eval().half()
     return model
